+++ src/presentation/controllers/signUp/signup.spec.ts
@@ -1,8 +1,7 @@
-
 import { SignUpController } from './signup'
 import { HttpRequest, AddAccount, AccountModel, AddAccountModel, Validation } from './signup-protocols'
 import { ServerError, MissingParamError } from '../../errors'
-import { badRequest } from '../../helpers/http-helper'
+import { badRequest } from '../../helpers/http/http-helper'
 
 const makeFakeRequest = (): HttpRequest => ({
   body: {
