+++ src/presentation/helpers/http/http-helper.ts
@@ -1,5 +1,5 @@
-import { HttpResponse } from '../protocols/http'
-import { ServerError, UnauthorizedError } from '../errors'
+import { HttpResponse } from '../../protocols/http'
+import { ServerError, UnauthorizedError } from '../../errors'
 
 export const badRequest = (error: Error): HttpResponse => ({
   statusCode: 400,
